added method getPWM()
Thank you for this great library!!

I added a method getPWM() to be able to access the current PWM values in private array pwmbuffer.

In a recent project I needed to calculate the **new** PWM value from the **current** PWM value. Instead of doing separate book-keeping in my sketch, I modified the library to be able to access the current values which have to be stored anyways.

This commit should not break any existing code out there, as it just adds new functionality.